src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dnspython==2.7.0
feedparser==6.0.11
fonttools==4.58.5
frozenlist==1.7.0
gnews==0.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiteconnect==5.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pandas_ta==0.3.14b0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyngrok==7.2.12
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
ruff==0.12.2
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tornado==6.5.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
zope.interface==7.2
